+++ examples/src/bin/teapot.rs
@@ -38,12 +38,12 @@ fn main() {
     println!("Using device: {} (type: {:?})", physical.name(), physical.ty());
 
     let mut events_loop = winit::EventsLoop::new();
-    let window = winit::WindowBuilder::new().build_vk_surface(&events_loop, instance.clone()).unwrap();
+    let surface = winit::WindowBuilder::new().build_vk_surface(&events_loop, instance.clone()).unwrap();
 
     let mut dimensions;
 
     let queue = physical.queue_families().find(|&q| q.supports_graphics() &&
-                                                   window.surface().is_supported(q).unwrap_or(false))
+                                                   surface.is_supported(q).unwrap_or(false))
                                                 .expect("couldn't find a graphical queue family");
 
     let device_ext = vulkano::device::DeviceExtensions {
@@ -57,7 +57,7 @@ fn main() {
     let queue = queues.next().unwrap();
 
     let (mut swapchain, mut images) = {
-        let caps = window.surface().capabilities(physical).expect("failed to get surface capabilities");
+        let caps = surface.capabilities(physical).expect("failed to get surface capabilities");
 
         dimensions = caps.current_extent.unwrap_or([1024, 768]);
 
@@ -65,7 +65,7 @@ fn main() {
         let format = caps.supported_formats[0].0;
         let alpha = caps.supported_composite_alpha.iter().next().unwrap();
 
-        vulkano::swapchain::Swapchain::new(device.clone(), window.surface().clone(), caps.min_image_count, format, dimensions, 1,
+        vulkano::swapchain::Swapchain::new(device.clone(), surface.clone(), caps.min_image_count, format, dimensions, 1,
                                            usage, &queue, vulkano::swapchain::SurfaceTransform::Identity,
                                            alpha,
                                            vulkano::swapchain::PresentMode::Fifo, true, None).expect("failed to create swapchain")
@@ -143,7 +143,7 @@ fn main() {
 
         if recreate_swapchain {
 
-        dimensions = window.surface().capabilities(physical)
+        dimensions = surface.capabilities(physical)
             .expect("failed to get surface capabilities")
             .current_extent.unwrap_or([1024, 768]);